improving the angular code
