+++ ArduinoFrontend/src/app/app.module.ts
@@ -12,25 +12,6 @@ import { CodeEditorComponent } from './code-editor/code-editor.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 // import { PathLocationStrategy, LocationStrategy } from '@angular/common';
 import { HashLocationStrategy, LocationStrategy } from '@angular/common';
-
-import {
-  MatDialogModule,
-  MatInputModule,
-  MatButtonModule,
-  MatTableModule,
-  MatRadioModule,
-  MatDividerModule,
-  MatTabsModule,
-  MatSnackBarModule,
-  MatIconModule,
-  MatTooltipModule,
-  MatExpansionModule,
-  MatMenuModule,
-  MatSidenavModule,
-  MatSliderModule,
-  MatListModule,
-  MatStepperModule
-} from '@angular/material';
 import { ViewComponentInfoComponent } from './view-component-info/view-component-info.component';
 import { HttpClientModule } from '@angular/common/http';
 import { ExportfileComponent } from './exportfile/exportfile.component';
@@ -47,6 +28,10 @@ import { SaveProjectDialogComponent } from './simulator/save-project-dialog/save
 import { OptionModalComponent } from './alert/option-modal/option-modal.component';
 import { VersioningPanelComponent } from './versioning-panel/versioning-panel.component';
 import { CreateVariationDialogComponent } from './versioning-panel/create-variation-dialog/create-variation-dialog.component';
+import { SidePanelComponent } from './side-panel/side-panel.component';
+import { MaterialModule } from './common/material.module';
+import { DashboardModule } from './dashboard/dashboard.module';
+import { SharedModule } from './common/SharedModule.module';
 
 /**
  * Monaco OnLoad Function
@@ -67,16 +52,15 @@ const monacoConfig: NgxMonacoEditorConfig = {
 @NgModule({
   declarations: [
     AppComponent,
-    DashboardComponent,
     SimulatorComponent,
     CodeEditorComponent,
     ViewComponentInfoComponent,
     ExportfileComponent,
     ComponentlistComponent,
     FrontPageComponent,
-    GalleryComponent,
+    // GalleryComponent,
     ViewProjectComponent,
-    HeaderComponent,
+    // HeaderComponent,
     AlertModalComponent,
     ConfirmModalComponent,
     OptionModalComponent,
@@ -85,6 +69,7 @@ const monacoConfig: NgxMonacoEditorConfig = {
     SaveProjectDialogComponent,
     VersioningPanelComponent,
     CreateVariationDialogComponent,
+    
   ],
   imports: [
     BrowserModule,
@@ -92,24 +77,10 @@ const monacoConfig: NgxMonacoEditorConfig = {
     FormsModule,
     MonacoEditorModule.forRoot(monacoConfig),
     BrowserAnimationsModule,
-    MatDialogModule,
-    MatRadioModule,
-    MatDividerModule,
-    MatInputModule,
-    MatButtonModule,
-    MatTableModule,
+    MaterialModule,
     HttpClientModule,
-    MatIconModule,
-    MatTabsModule,
-    MatTooltipModule,
-    MatSnackBarModule,
-    MatExpansionModule,
-    MatSidenavModule,
-    MatSliderModule,
-    MatListModule,
-    MatStepperModule,
     ReactiveFormsModule,
-    MatMenuModule
+    SharedModule   //shared module for Header component
   ],
   // providers: [{provide: LocationStrategy, useClass: PathLocationStrategy}],
   providers: [{ provide: LocationStrategy, useClass: HashLocationStrategy }],
@@ -129,6 +100,6 @@ const monacoConfig: NgxMonacoEditorConfig = {
   schemas: [
     CUSTOM_ELEMENTS_SCHEMA
   ],
-  exports: [HeaderComponent]
+  exports: [AppComponent]
 })
 export class AppModule { }
