+++ ArduinoFrontend/src/app/common/SharedModule.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { HeaderComponent } from '../header/header.component';
+import { RouterModule } from '@angular/router';
+import { MaterialModule } from './material.module';
+
+@NgModule({
+    imports: [
+        CommonModule,
+        RouterModule,
+        MaterialModule
+     ],
+    declarations: [
+        HeaderComponent
+    ],
+    exports: [
+        HeaderComponent
+    ]
+})
+export class SharedModule {}
