+++ ArduinoFrontend/src/app/dashboard/dashboard-routing.module.ts
@@ -1,0 +1,34 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { GalleryComponent } from '../gallery/gallery.component';
+import { MainPageComponent } from '../main-page/main-page.component';
+import { DashboardComponent } from './dashboard.component';
+
+
+const routes: Routes = [
+    {
+        path: '',
+        component: DashboardComponent,
+        children: [
+            {
+                path: '',
+                component:MainPageComponent
+
+            },
+            {
+                path: 'gallery',
+                component: GalleryComponent
+            },
+            {
+                path :'mainpage/:id',
+                component:MainPageComponent
+            }
+        ]
+    },
+];
+
+@NgModule({
+    imports: [RouterModule.forChild(routes)],
+    exports: [RouterModule]
+})
+export class DashboardRoutingModule { }
