+++ ArduinoFrontend/src/app/dashboard/dashboard.component.css
@@ -1,152 +1,14 @@
-.card:hover {
-  background-color: #ddd;
-}
-
-.subbox {
-  box-sizing: border-box;
-  width: 95%;
-  height: auto;
-  background-color: white;
-  border-radius: 0.3125em;
-  margin-bottom: 15px;
-  box-shadow: 1px 1px 3px grey;
-  cursor: pointer;
-  margin-left: 2%;
-  justify-content: space-evenly;
-}
-
-.item-img {
-  min-height: 180px;
-  max-width: 100%;
-  max-height: 180px;
-}
-
-.subbox:hover {
-  box-shadow: 2px 2px 9px grey;
-}
-
-#openproject {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100vh;
-  background-color: rgba(0, 0, 0, 0.4);
-  overflow-y: scroll;
-}
-
-#innerdialog {
-  background-color: white;
-  position: relative;
-  color: black;
-  margin: 1em 10% 1em 10%;
-  border-radius: 5px;
-  box-sizing: border-box;
-  min-width: 80%;
-}
-
-button[type="submit"] {
-  color: white;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  float: left;
-  font-weight: bold;
-  border-top: 1px solid #007bff;
-  border-bottom: 1px solid #007bff;
-  border-top-left-radius: 0;
-  border-bottom-left-radius: 0;
-  height: 100%;
-}
-
-#Search {
-  width: 100%;
-  border-top-right-radius: 0;
-  border-bottom-right-radius: 0;
-}
-
-.act-buttons {
-  max-width: 100%;
-}
-
-.act-buttons button {
-  margin-left: 10px;
-  margin-bottom: 10px;
-}
-
-.share-btns {
-  margin-top: 1.5em;
-  margin-bottom: 1em;
-  width: 100%;
-}
-
-.share-btns button {
-  margin-left: 15px;
-}
-
-.act-desc {
-  margin: 15px;
-  font-family: cursive;
-}
-
-@media only screen and (max-width: 480px) {
-  .share-btns button {
-    margin-left: 5px;
-  }
-}
-/* .example-container {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-} 
-*/
-.role-container {
-  width: 300px;
-  margin-bottom: 20px;
-  box-shadow: 2px 0px #495057;
-}
 
 .example-container {
   min-height: 876px;
   border: 1px solid rgba(0, 0, 0, 0.12);
 }
-.example-sidenav-content {
-  /* display: flex; */
-  /* height: 100% !important; */
-  /* align-items: center;
-  justify-content: center; */
-}
-.example-sidenav {
-  min-width: 200px;
-  user-select: none;
-}
-.full-width {
-  width: 100%;
-}
-.menu-button {
-  transition: 300ms ease-in-out;
-  transform: rotate(0deg);
-}
-.menu-button.rotated {
-  transform: rotate(180deg);
-}
-.submenu {
-  display: none;
-  overflow-y: hidden;
-  transition: transform 300ms ease;
-  transform: scaleY(0);
-  transform-origin: top;
-  padding-left: 30px;
-}
-.submenu.expanded {
-  display: block;
-  transform: scaleY(1);
-}
-.custom-menu {
-  margin-right: 1.3rem;
-  margin-left: 8px;
+
+.no-search {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 1rem;
 }
 .maincontent {
   display: flex;
@@ -154,6 +16,13 @@ button[type="submit"] {
   flex-direction: column;
   align-items: center;
 }
-.circuitBtn{
+.circuitBtn {
   width: 296px;
+}
+
+mat-chip {
+  max-width: 200px;
+  padding: 9px;
+  border-radius: 14px;
+  padding: 11px;
 }
