+++ ArduinoFrontend/src/app/header/header.component.html
@@ -15,7 +15,7 @@
     <a mat-stroked-button [href]="window.location.protocol + '//' + window.location.hostname"
       style="margin-right: 10px;">Home</a>
     <a mat-stroked-button [routerLink]="['/dashboard']" style="margin-right: 10px;">Dashboard</a>
-    <a mat-stroked-button [routerLink]="['/gallery']" style="margin-right: 10px;">Gallery</a>
+    <a mat-stroked-button [routerLink]="['/dashboard/gallery']" style="margin-right: 10px;">Gallery</a>
     <a mat-stroked-button [routerLink]="['/simulator']" style="margin-right: 10px">Editor</a>
     <a mat-stroked-button (click)="Login()" *ngIf="!username">Login</a>
   </div>
@@ -48,7 +48,7 @@
       </div>
     </div>
   </div>-->
-      <div class="dropdown show mr-2" *ngIf="token" style="margin-top: 4px;">
+      <div class="dropdown show mr-2" *ngIf="token" style="margin-top: 9px;">
         <!-- <img class="dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-toggle="dropdown"
           aria-haspopup="true" aria-expanded="false" src="./assets/images/login.svg" width="30" height="30"
           class="d-inline-block align-center" alt="Login Image"> -->
