+++ ArduinoFrontend/src/app/header/header.component.ts
@@ -1,8 +1,9 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, OnDestroy } from '@angular/core';
 import { Login } from '../Libs/Login';
 import { ApiService } from '../api.service';
 import { environment } from 'src/environments/environment';
 import { ActivatedRoute } from '@angular/router';
+import { Subscription } from 'rxjs';
 
 /**
  * Class For Header Component (DO Eager Loading)
@@ -12,7 +13,7 @@ import { ActivatedRoute } from '@angular/router';
   templateUrl: './header.component.html',
   styleUrls: ['./header.component.css']
 })
-export class HeaderComponent implements OnInit {
+export class HeaderComponent implements OnInit, OnDestroy {
   /**
    * Login Token
    */
@@ -46,6 +47,11 @@ export class HeaderComponent implements OnInit {
    * Constructor for Header
    * @param api API Service
    */
+
+  /**
+   * getting subscription of user;
+   */
+  subscription = new Subscription();
   constructor(
     private api: ApiService,
     private aroute: ActivatedRoute) { }
@@ -53,7 +59,8 @@ export class HeaderComponent implements OnInit {
    * On Init
    */
   ngOnInit() {
-    this.userInfo();
+    this.getUserInfo();
+    this.stateChangeUserInfo();
     // Initializing window
     this.window = window;
   }
@@ -73,8 +80,20 @@ export class HeaderComponent implements OnInit {
   /**
    * Getting User Information.
    */
-  userInfo() {
+  stateChangeUserInfo() {
     // Get Login Token
+    this.subscription = this.api.isAuthenticated.subscribe((res: boolean) => {
+      if (res == true) {
+        this.getUserInfo();
+      } else {
+        this.username = undefined;
+      }
+
+    })
+
+  }
+
+  getUserInfo() {
     this.token = Login.getToken();
     // If token is available then get username
     if (this.token) {
@@ -89,5 +108,12 @@ export class HeaderComponent implements OnInit {
         // }
       });
     }
+  }
+  /**
+   * on destroy when page is dead then call
+   */
+  ngOnDestroy() {
+    // destroy  auth subject.
+    this.subscription.unsubscribe();
   }
 }
