+++ ArduinoFrontend/src/app/simulator/simulator.component.html
@@ -49,18 +49,17 @@
       <a mat-raised-button
         (click)="HandleRouter(RouteToFunction(window.location.protocol + '//' + window.location.hostname, true))"
         class="navBarItem">Home</a>
-      <a mat-raised-button (click)="HandleRouter(RouteToFunction('/dashboard'))"
-        class="navBarItem">Dashboard</a>
-      <a mat-raised-button (click)="HandleRouter(RouteToFunction('/gallery'))" class="navBarItem">Gallery</a>
+      <a mat-raised-button (click)="HandleRouter(RouteToFunction('/dashboard'))" class="navBarItem">Dashboard</a>
+      <a mat-raised-button (click)="HandleRouter(RouteToFunction('/dashboard/gallery'))" class="navBarItem">Gallery</a>
       <a mat-raised-button (click)="HandleRouter(RouteToSimulator.bind(this))" class="navBarItem">Editor</a>
       <a mat-raised-button *ngIf="!token" (click)="HandleRouter(Login)">Login</a>
-      <div class="dropdown show" style="margin-left: 16px;" *ngIf="token">
+      <div class="dropdown show" style="margin-left: 10px;margin-top:4px;" *ngIf="token">
         <!-- <img class="dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-toggle="dropdown"
           aria-haspopup="true" aria-expanded="false" src="./assets/images/login.svg" width="30" height="30"
           class="d-inline-block align-center" alt="Login Image"> -->
-          <div class="circles" id="dropdownMenuLink" data-toggle="dropdown">
-            {{ username.charAt(0).toUpperCase() }}
-          </div>
+        <div class="circles" id="dropdownMenuLink" data-toggle="dropdown">
+          {{ username.charAt(0).toUpperCase() }}
+        </div>
         <div class="dropdown-menu dropdown-menu-right" style="background-color: whitesmoke;"
           aria-labelledby="dropdownMenuLink">
           <a *ngIf="username" class="dropdown-item"><i class="fas fa-user mr-1"></i> {{username}}</a>
